\name{ToPercentile}
\alias{ToPercentile}
\title{Convert Values to Percentiles}
\usage{
ToPercentile(x)
}
\arguments{
  \item{x}{a numeric vector}
}
\description{
This function converts values in a numeric vector to their corresponding percentiles.
}
\details{
The function takes one argument: \code{x}, which is a numeric vector.

The function first creates an empirical cumulative distribution function (ECDF) using the \code{ecdf()} function.

Then, it uses the ECDF to calculate the percentiles of the values in \code{x} by calling the ECDF function with \code{x} as the input.

The function returns a numeric vector of the same length as \code{x}, where each value represents the percentile of the corresponding value in \code{x}.

Note that the percentiles are calculated based on the empirical distribution of the values in \code{x}.
}
\examples{
# Convert values to percentiles
x <- c(2, 4, 6, 8, 10)
percentiles <- ToPercentile(x)
percentiles
}
